+++ packages/egenie-common/src/print/index.ts
@@ -1,4 +1,5 @@
-export { printHelper, formatBarcodeData } from './printHelper';
+export { printHelper } from './printHelper';
 export { printWayBill } from './printWayBill';
 export { CustomPrintModal, getCustomPrintParam, getCustomPrintParamByDefaultTemplate } from './customPrint';
 export { getSensitiveData } from './getPrivacyData';
+export { formatBarcodeData } from './utils';