+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # [0.0.0](http://192.168.200.111:9980/egFrontend/egenie-utils/compare/v0.7.5...v0.0.0) (2021-12-03)
 
+### Features
+
+- 删除无用的 readme ([15bbb54](http://192.168.200.111:9980/egFrontend/egenie-utils/commits/15bbb5486f019ce631bf307f070be44e01f25cd4))
+- **egenie-common:** 打印辅助函数的测试 ([567ee5d](http://192.168.200.111:9980/egFrontend/egenie-utils/commits/567ee5df1efc01dec7d215b53449966c9bc79f16))
+
 ## [0.7.5](http://192.168.200.111:9980/egFrontend/egenie-utils/compare/v0.7.4...v0.7.5) (2021-12-02)
 
 ### Features